Create Snes::GameGenie
Just a bit of pre-emptive separation. I intend to allow for the parsing
of other code types in the future.
